+++ companion plugins/tbc-starter-theme/functions.php
@@ -13,7 +13,7 @@ if (!defined('ABSPATH')) {
 /**
  * Theme Constants
  */
-define('FLUENT_STARTER_VERSION', '1.0.49');
+define('FLUENT_STARTER_VERSION', '1.0.50');
 define('FLUENT_STARTER_DIR', get_template_directory());
 define('FLUENT_STARTER_URI', get_template_directory_uri());